requirements.txt
# BFS-Prover MCP Server Dependencies

# MCP server framework
mcp>=1.0.0

# Model inference
llama-cpp-python>=0.3.0

# Data validation
pydantic>=2.0.0

# System utilities
psutil>=5.9.0
